feat(metrics): add Prometheus metrics support for beacon client
* Implement core beacon metrics (head slot, epochs, active validators, reorgs, deposits)

* Expose configurable metrics address and port in the beacon node CLI

* Integrate with Grafana dashboard via `beaconMetrics.json`

* Update ream_launcher.star + Kurtosis integration

Closes #1526
