+++ bin/ream/src/cli/beacon_node.rs
@@ -8,9 +8,9 @@ use ream_p2p::bootnodes::Bootnodes;
 use url::Url;
 
 use crate::cli::constants::{
-    DEFAULT_DISABLE_DISCOVERY, DEFAULT_DISCOVERY_PORT, DEFAULT_HTTP_ADDRESS,
-    DEFAULT_HTTP_ALLOW_ORIGIN, DEFAULT_HTTP_PORT, DEFAULT_NETWORK, DEFAULT_SOCKET_ADDRESS,
-    DEFAULT_SOCKET_PORT,
+    DEFAULT_BEACON_METRICS_ADDRESS, DEFAULT_BEACON_METRICS_PORT, DEFAULT_DISABLE_DISCOVERY,
+    DEFAULT_DISCOVERY_PORT, DEFAULT_HTTP_ADDRESS, DEFAULT_HTTP_ALLOW_ORIGIN, DEFAULT_HTTP_PORT,
+    DEFAULT_METRICS_ENABLED, DEFAULT_NETWORK, DEFAULT_SOCKET_ADDRESS, DEFAULT_SOCKET_PORT,
 };
 
 #[derive(Debug, Parser)]
@@ -89,6 +89,15 @@ pub struct BeaconNodeConfig {
         help = "Number of epochs to retain blob sidecars. Defaults to network spec value (4096 epochs for mainnet, ~18 days)"
     )]
     pub blob_retention_epochs: Option<u64>,
+
+    #[arg(long = "metrics", help = "Enable metrics", default_value_t = DEFAULT_METRICS_ENABLED)]
+    pub enable_metrics: bool,
+
+    #[arg(long, help = "Set metrics address", default_value_t = DEFAULT_BEACON_METRICS_ADDRESS)]
+    pub metrics_address: IpAddr,
+
+    #[arg(long, help = "Set metrics port", default_value_t = DEFAULT_BEACON_METRICS_PORT)]
+    pub metrics_port: u16,
 }
 
 impl From<BeaconNodeConfig> for ManagerConfig {
