+++ bin/ream/src/cli/constants.rs
@@ -8,6 +8,8 @@ pub const DEFAULT_HTTP_ADDRESS: IpAddr = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1))
 pub const DEFAULT_HTTP_ALLOW_ORIGIN: bool = false;
 pub const DEFAULT_HTTP_PORT: u16 = 5052;
 pub const DEFAULT_KEY_MANAGER_HTTP_PORT: u16 = 8008;
+pub const DEFAULT_BEACON_METRICS_ADDRESS: IpAddr = IpAddr::V4(Ipv4Addr::UNSPECIFIED);
+pub const DEFAULT_BEACON_METRICS_PORT: u16 = 8008;
 pub const DEFAULT_METRICS_ENABLED: bool = false;
 pub const DEFAULT_METRICS_ADDRESS: IpAddr = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
 pub const DEFAULT_METRICS_PORT: u16 = 8080;
