+++ crates/common/consensus/beacon/Cargo.toml
@@ -43,6 +43,7 @@ ream-consensus-misc.workspace = true
 ream-execution-engine.workspace = true
 ream-execution-rpc-types.workspace = true
 ream-merkle.workspace = true
+ream-metrics.workspace = true
 ream-network-spec.workspace = true
 
 [dev-dependencies]
