+++ crates/common/consensus/beacon/src/matrix_entry.rs
@@ -3,6 +3,10 @@ use ream_consensus_misc::{
     constants::beacon::CELLS_PER_EXT_BLOB, polynomial_commitments::kzg_proof::KZGProof,
 };
 use ream_execution_rpc_types::get_blobs::Blob;
+use ream_metrics::{
+    BEACON_DATA_AVAILABILITY_RECONSTRUCTED_COLUMNS_TOTAL,
+    BEACON_DATA_AVAILABILITY_RECONSTRUCTION_TIME_SECONDS,
+};
 use rust_eth_kzg::{Cell as KZGCell, DASContext, KZGProof as Proof};
 use ssz_types::FixedVector;
 
@@ -40,7 +44,9 @@ pub fn recover_matrix(
     blob_count: u64,
     das_context: &DASContext,
 ) -> Result<Vec<MatrixEntry>> {
+    let _timer = BEACON_DATA_AVAILABILITY_RECONSTRUCTION_TIME_SECONDS.start_timer();
     let mut matrix = Vec::new();
+    let mut reconstructed_count: u64 = 0;
 
     for blob_index in 0..blob_count {
         let (cell_indices, cells): (Vec<u64>, Vec<Cell>) = partial_matrix
@@ -49,8 +55,13 @@ pub fn recover_matrix(
             .map(|entry| (entry.column_index, entry.cell.clone()))
             .unzip();
 
+        let known_count = cell_indices.len() as u64;
+
         let (recovered_cells, recovered_proofs) =
             recover_cells_and_kzg_proofs(cell_indices, cells, das_context)?;
+
+        let newly_reconstructed = recovered_cells.len() as u64 - known_count;
+        reconstructed_count += newly_reconstructed;
 
         for (cell_index, (cell, kzg_proof)) in recovered_cells
             .into_iter()
@@ -65,6 +76,8 @@ pub fn recover_matrix(
             });
         }
     }
+
+    BEACON_DATA_AVAILABILITY_RECONSTRUCTED_COLUMNS_TOTAL.inc_by(reconstructed_count);
 
     Ok(matrix)
 }
