+++ crates/common/fork_choice/beacon/Cargo.toml
@@ -40,6 +40,7 @@ ream-consensus-lean.workspace = true
 ream-consensus-misc.workspace = true
 ream-execution-engine.workspace = true
 ream-execution-rpc-types.workspace = true
+ream-metrics.workspace = true
 ream-network-spec.workspace = true
 ream-operation-pool.workspace = true
 ream-polynomial-commitments.workspace = true
