+++ crates/common/polynomial_commitments/Cargo.toml
@@ -35,6 +35,7 @@ ream-bls.workspace = true
 ream-consensus-beacon.workspace = true
 ream-consensus-misc.workspace = true
 ream-execution-rpc-types.workspace = true
+ream-metrics.workspace = true
 
 [lints]
 workspace = true
